notes about possibility of storing all segments in one file and the page management thereof.  changed btree cursor to accept the root page number instead of the stream length.

+++ tests/combo.cs
@@ -28,7 +28,7 @@ namespace lsm_tests
 	{
 		public Func<IWrite> create_memory_segment;
 		public Func<Stream,ICursor,uint> create_btree_segment;
-		public Func<Stream,long,ICursor> open_btree_segment;
+		public Func<Stream,uint,ICursor> open_btree_segment;
 		public Func<ICursor,ICursor> create_living_cursor;
 		public Func<ICursor,ICursor,ICursor> create_multicursor;
